moved pdc/epdl/deol/extent_buffer related code to pdc.c and pdc.h
